Fix for 8.1.0 Alpha 3

+++ app/config/defaultBBCode.php
@@ -261,8 +261,8 @@ HANDLER,
         'handler' => <<<'HANDLER'
 $arr = [];
 $sec = $attrs['Def'] % 60;
-$min = ($attrs['Def'] / 60) % 60;
-$hours = ($attrs['Def'] / 3600) % 24;
+$min = (int) ($attrs['Def'] / 60) % 60;
+$hours = (int) ($attrs['Def'] / 3600) % 24;
 $days = (int) ($attrs['Def'] / 86400);
 if ($days > 0) {
     $arr[] = $days . __('After time d');
